# Adding support for efficiently merging video datasets