Adds support for efficiently merging video datasets using aggregations, as opposed to explicit iteration over the samples.

This efficient implementation is automatically used when `Dataset.merge_samples()` is called as long as no optional kwargs are passed that prevents its use.

### Benchmarking

The examples below can be reverted to the non-efficient implementation by passing `overwrite=False` to `merge_samples()`. Here are the benchmarking results for each example:

Image example
- This PR: `63.7 ms`
- Previous: `6.1 seconds`

Video example:
- This PR: `2.6 seconds`
- Previous: `34.0 seconds`

### Example: merging an image dataset

```py
import os

import eta.core.utils as etau

import fiftyone as fo
import fiftyone.zoo as foz
from fiftyone import ViewField as F

IMAGES_DIR = "/tmp/quickstart"

dataset = foz.load_zoo_dataset("quickstart").clone()

dataset1 = dataset.clone()
dataset2 = dataset.clone()
dataset3 = dataset.clone()

# Give `dataset2` new field names
dataset2.rename_sample_field("ground_truth", "gt")

# Give `dataset3` new image locations
inpaths = dataset3.values("filepath")
outpaths = [os.path.join(IMAGES_DIR, os.path.basename(f)) for f in inpaths]
for inpath, outpath in zip(inpaths, outpaths):
    etau.copy_file(inpath, outpath)

dataset3.set_values("filepath", outpaths)

# Only merge `person` labels
view2 = dataset2.filter_labels("gt", F("label") == "person")
view3 = dataset3.filter_labels("ground_truth", F("label") == "person")

# Perform the merges
dataset.merge_samples(view2)
dataset.merge_samples(view3)

# Check that schema is correct
print(dataset)

#
# `dataset` should contain:
#   `view2` counts in its `gt` field
#   `dataset1` + `view3` counts in its `ground_truth` field
#

print(dataset.count_values("ground_truth.detections.label"))
print(dataset1.count_values("ground_truth.detections.label"))
print(view3.count_values("ground_truth.detections.label"))

print(dataset.count_values("gt.detections.label"))
print(view2.count_values("gt.detections.label"))

# Double-check everything is working in App
session = fo.launch_app(dataset)
```

### Example: merging a video dataset

```py
import os

import eta.core.utils as etau

import fiftyone as fo
import fiftyone.zoo as foz
from fiftyone import ViewField as F

VIDEOS_DIR = "/tmp/quickstart-video"

dataset = foz.load_zoo_dataset("quickstart-video").clone()
dataset.rename_frame_field("ground_truth_detections", "ground_truth")

dataset1 = dataset.clone()
dataset2 = dataset.clone()
dataset3 = dataset.clone()

# Give `dataset2` new field names
dataset2.rename_frame_field("ground_truth", "gt")

# Give `dataset3` new video locations
inpaths = dataset3.values("filepath")
outpaths = [os.path.join(VIDEOS_DIR, os.path.basename(f)) for f in inpaths]
for inpath, outpath in zip(inpaths, outpaths):
    etau.copy_file(inpath, outpath)

dataset3.set_values("filepath", outpaths)

# Only merge `vehicle` labels
view2 = dataset2.filter_labels("frames.gt", F("label") == "vehicle")
view3 = dataset3.filter_labels("frames.ground_truth", F("label") == "vehicle")

dataset.merge_samples(view2)
dataset.merge_samples(view3)

# Check that schema is correct
print(dataset)

#
# `dataset` should contain:
#   `view2` counts in its `gt` field
#   `dataset1` + `view3` counts in its `ground_truth` field
#

print(dataset.count_values("frames.ground_truth.detections.label"))
print(dataset1.count_values("frames.ground_truth.detections.label"))
print(view3.count_values("frames.ground_truth.detections.label"))

print(dataset.count_values("frames.gt.detections.label"))
print(view2.count_values("frames.gt.detections.label"))

# Double-check everything is working in App
session = fo.launch_app(dataset)
```